LayoutTests/http/tests/security/xss-DENIED-window-open-parent-expected.txt

CONSOLE MESSAGE: line 3: Blocked a frame with origin "http://localhost:8080" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
This test passes if there is no alert dialog.
